Renamed tje two project dialog files to something a bit shorter
