+++ nw/gui/__init__.py
@@ -15,8 +15,8 @@ from nw.gui.itemeditor import GuiItemEditor
 from nw.gui.mainmenu import GuiMainMenu
 from nw.gui.outline import GuiProjectOutline
 from nw.gui.preferences import GuiPreferences
-from nw.gui.projectload import GuiProjectLoad
-from nw.gui.projectsettings import GuiProjectSettings
+from nw.gui.projload import GuiProjectLoad
+from nw.gui.projsettings import GuiProjectSettings
 from nw.gui.sessionlog import GuiSessionLogView
 from nw.gui.statusbar import GuiMainStatus
 from nw.gui.theme import GuiIcons
