+++ nw/gui/projsettings.py
